GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - get-current-user-actions (refs/gitbutler/get-current-user-actions)
   branch head: e176ccbb5f60e3e8432f3eb3fa0fc2cb2236f8e5
 - update-login-link (refs/gitbutler/update-login-link)
   branch head: ab5804809aec3c1724453dade02e31750798fdd5
 - init
 (refs/gitbutler/init-)
   branch head: 8f4d3e24c5bdfc7da7ddc0a03e834941d601aeb2
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch